PFUAggDatabase data frame column names

Aggregation data frame column names

Usage

aggregation_df_cols

Format

A string list with 9 entries.

product_aggregation

The name of the metadata column that tells about product aggregation. "Product.Aggregation"

industry_aggregation

The name of the metadata column that tells about industry aggregation. "Industry.Aggregation"

specified

The value that indicates products or industries remain is specified. "Specified"

despecified

The value that indicates products or industries have been despecified and aggregated. "Despecified"

ungrouped

The value that indicates products or industries have not been grouped. "Ungrouped"

grouped

The value that indicates products or industries have been grouped. "Grouped"

chopped_mat

The value that indicates which matrix has been chopped. "Chopped.Mat"

chopped_var

The value that indicates the chopping product or industry. "Chop.Var"

product_sector

The column containing values for chopped_var. Recca::aggregate_cols$product_sector.

Details

A string list containing names of column names and values for aggregation data frames.

Examples

aggregation_df_cols
#> $product_aggregation
#> [1] "Product.aggregation"
#> 
#> $industry_aggregation
#> [1] "Industry.aggregation"
#> 
#> $specified
#> [1] "Specified"
#> 
#> $despecified
#> [1] "Despecified"
#> 
#> $ungrouped
#> [1] "Ungrouped"
#> 
#> $grouped
#> [1] "Grouped"
#> 
#> $chopped_mat
#> [1] "Chopped.mat"
#> 
#> $chopped_var
#> [1] "Chopped.var"
#> 
#> $product_sector
#> [1] "Product.Industry.Sector"
#>