Calculate efficiencies from an aggregates data frame

This function calculates efficiencies (etas) from the aggregates data frame. Aggregate energy and exergy are retained.

Usage

calc_agg_etas(
  .aggregates,
  stage_colname = PFUAggDatabase::sea_cols$stage_colname,
  ex_colname = PFUAggDatabase::sea_cols$ex_colname,
  agg_by_colname = PFUAggDatabase::sea_cols$agg_by_colname,
  e_product_colname = PFUAggDatabase::sea_cols$e_product_colname,
  gross_net_colname = PFUAggDatabase::sea_cols$gross_net_colname,
  sector_colname = PFUAggDatabase::sea_cols$sector_colname,
  eta_pf_colname = PFUAggDatabase::efficiency_cols$eta_pf,
  eta_fu_colname = PFUAggDatabase::efficiency_cols$eta_fu,
  eta_pu_colname = PFUAggDatabase::efficiency_cols$eta_pu,
  gross = PFUAggDatabase::gross_net_metadata$gross,
  net = PFUAggDatabase::gross_net_metadata$net,
  total_value = PFUAggDatabase::agg_metadata$total_value,
  primary = IEATools::all_stages$primary,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  flow = IEATools::iea_cols$flow,
  tar_group = "tar_group"
)

Arguments

.aggregates

The data frame from which efficiencies are to be calculated. This data frame should be the output of the PSUT_Re_all_St_pfu or PSUT_Re_all_St_pfu_by_country targets.

stage_colname, ex_colname, agg_by_colname, e_product_colname, gross_net_colname, sector_colname

See PFUAggDatabase::sea_cols.

eta_pf_colname, eta_fu_colname, eta_pu_colname

See PFUAggDatabase::efficiency_cols.

gross, net

See PFUAggDatabase::gross_net_metadata

total_value

See PFUAggDatabase::agg_metadata.

primary, final, useful, country, year, method, energy_type, flow

See IEATools::iea_cols.

tar_group

The name of the targets grouping column in .aggregates. This column is deleted on output.

Value

A data frame of aggregates and aggregate efficiencies.

Details

This function removes the tar_group column.