Calculate primary aggregates for PSUT data

This function routes to Recca::primary_aggregates().

Usage

calculate_primary_aggregates(
  .psut_data,
  countries,
  years,
  p_industries,
  piece = "noun",
  notation = list(RCLabels::bracket_notation, RCLabels::arrow_notation),
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list
)

Arguments

.psut_data

The data for which primary aggregates are to be calculated.

countries

The countries for which primary aggregates are to be calculated.

years

The years for which primary aggregates are to be calculated.

p_industries

A string vector of industries that count as "primary".

piece

The piece to be aggregated. Default is "noun".

notation

The assumed notation for the labels. Default is list(RCLabels::bracket_notation, RCLabels::arrow_notation).

pattern_type

The type of matching to be used for primary industry names. Default is "leading".

prepositions

The expected propositions in row and column labels. Default is RCLabels::prepositions_list.

Value

A version of .psut_data with additional column for primary aggregate data.