Create an aggregation map for continent aggregation

An aggregation map is a named list. Names are the continents. List items are 3-letter country codes.

continent_aggregation_map(
  exemplar_table_path,
  exemplar_table_tab = "exemplar_table",
  region_code = "Region.code",
  continent = "Continent",
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year
)

Arguments

exemplar_table_path

The path to the exemplar table

exemplar_table_tab

The name of the tab in the file at exemplar_table_path where the region information exists. Default is "exemplar_table".

region_code

The name of a column containing region codes. Default is "Region.code".

continent

The name of the continent column. Default is "Continent".

country

The name of the country column in the outgoing data frame. Default is IEATools::iea_cols$country.

year

The name of the year column in the outgoing data frame. Default is IEATools::iea_cols$year.

Value

A continent aggregation map.

Details

The exemplar table is assumed to be an Excel file with the following columns: region_code and years (as numbers). The body of the table should contain 3-letter codes of countries. The exemplar table is assumed to be on the exemplar_table_tab tab of the Excel file.