Calculate PFU aggregations and efficiencies

Calculate PFU aggregations and efficiencies

Usage

efficiency_pipeline(
  .psut_data,
  p_industries,
  fd_sectors,
  piece = "noun",
  notation = list(RCLabels::bracket_notation, RCLabels::arrow_notation),
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  country = Recca::psut_cols$country,
  method = Recca::psut_cols$method,
  energy_type = Recca::psut_cols$energy_type,
  ieamw = PFUDatabase::ieamw_cols$ieamw,
  year = Recca::psut_cols$year,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  gross_net = Recca::efficiency_cols$gross_net,
  last_stage = Recca::psut_cols$last_stage,
  primary = IEATools::all_stages$primary,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  chopped_mat = PFUAggDatabase::aggregation_df_cols$chopped_mat,
  chopped_var = PFUAggDatabase::aggregation_df_cols$chopped_var,
  none = PFUAggDatabase::agg_metadata$none,
  ex_p = Recca::aggregate_cols$aggregate_primary,
  ex_f = IEATools::aggregate_cols$aggregate_final,
  ex_u = IEATools::aggregate_cols$aggregate_useful,
  ex_fd_gross = Recca::aggregate_cols$gross_aggregate_demand,
  ex_fd_net = Recca::aggregate_cols$net_aggregate_demand,
  ex_fd = Recca::aggregate_cols$aggregate_demand,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu,
  tol = 1e-06,
  .primary_aggs_ok = ".primary_aggs_OK",
  .final_aggs_ok = ".final_aggs_OK",
  .useful_aggs_ok = ".useful_aggs_OK"
)

Arguments

.psut_data

A data frame of PSUT data or a slice (row) of the data frame.

p_industries

A string vector of primary industries.

fd_sectors

A string vector of final demand sectors.

piece

The piece of the labels used for matching. Default is "noun".

notation

The notation used for row and column labels. Default is list(RCLabels::bracket_notation, RCLabels::arrow_notation).

pattern_type

The pattern type to be used for row and column matching. Default is "exact".

prepositions

A list of prepositions for row and column labels. Default is RCLabels::prepositions_list.

R, U, V, Y, r_eiou, U_eiou, U_feed, S_units, country, method, energy_type, year, last_stage

See Recca::psut_cols.

ieamw

See PFUDatabase::ieamw_cols.

gross, net, gross_net

See Recca::efficiency_cols.

primary, final, useful

See IEATools::all_stages.

chopped_mat, chopped_var

Column names identifying chopped matrices and variables. See PFUAggDatabase::aggregation_df_cols

none

A string signifying no aggregation is present. Default is PFUAggDatabase::agg_metadata$none.

ex_p, ex_fd_gross, ex_fd_net, ex_fd

See Recca::aggregate_cols.

ex_f, ex_u

See IEATools::aggregate_cols.

eta_pf, eta_fu, eta_pu

See Recca::efficiency_cols.

tol

The amount by which aggregate primary, final, and useful values can be different before an error is thrown. Default is 1e-6.

.primary_aggs_ok, .final_aggs_ok, .useful_aggs_ok

Names of columns used internally to assess whether primary, final, and useful aggregations are all same.

Value

A data frame of metadata columns and efficiencies