Create a targets workflow

Arguments to this function specify the details of a targets workflow to be executed.

get_pipeline(
  countries = NULL,
  psut_release,
  psut_releases_folder,
  exemplar_table_path,
  world_agg_map
)

Arguments

countries

A string vector of 3-letter country codes. Default is NULL, meaning all countries should be analyzed.

psut_release

The release we'll use from psut_releases_folder.

psut_releases_folder

The path to the pins archive of PSUT releases.

exemplar_table_path

The path to the examplar table.

world_agg_map

The aggregation map to aggregate from continents to the world.

Value

A list of tar_targets to be executed in a workflow.

Details

The exemplar table is assumed to be an Excel file with the following columns: "Region.code" and years (as numbers). The body of the table should contain 3-letter codes of countries. The exemplar table is assumed to be on the "exemplar_table" tab of the Excel file.