Pivot and write aggregate efficiencies to an Excel file

The incoming data frame is expected to contain a year column as well as primary, final, useful, eta_pf, eta_fu, and eta_pu columns.

Usage

write_agg_etas_xlsx(
  .agg_etas,
  path,
  aggs_tabname = PFUAggDatabase::output_file_info$agg_tabname,
  etas_tabname = PFUAggDatabase::output_file_info$eta_tabname,
  wide_by_year = TRUE,
  primary = IEATools::all_stages$primary,
  final = IEATools::all_stages$final,
  useful = IEATools::all_stages$useful,
  eta_pf_colname = PFUAggDatabase::efficiency_cols$eta_pf,
  eta_fu_colname = PFUAggDatabase::efficiency_cols$eta_fu,
  eta_pu_colname = PFUAggDatabase::efficiency_cols$eta_pu,
  year = IEATools::iea_cols$year,
  country = IEATools::iea_cols$country,
  quantity = IEATools::template_cols$quantity,
  .values = IEATools::template_cols$.values
)

Arguments

.agg_etas

A data frame created by the eta_Re_all_St_pfu target.

path

The path where the Excel file will be saved.

aggs_tabname, etas_tabname

See PFUAggDatabase::output_file_info.

wide_by_year

If TRUE (the default), the incoming data frame will be pivoted to be wide by years. If FALSE, data will be unchanged.

primary, final, useful

See IEATools::all_stages.

eta_pf_colname, eta_fu_colname, eta_pu_colname

See PFUAggDatabase::efficiency_cols.

year, country

See IEATools::iea_cols.

quantity, .values

See IEATools::template_cols.

Value

TRUE if the file was written successfully.